+++ internal/config/capabilities.go
@@ -1,0 +1,5 @@
+package config
+
+type Capabilities struct {
+	AvailabilityZones []string
+}