+++ internal/config/config.go
@@ -1,9 +1,10 @@
 package config
 
 type Flags struct {
-	Server     Server
-	Identity   Identity
-	Kubernetes Kubernetes
+	Server       Server
+	Identity     Identity
+	Kubernetes   Kubernetes
+	Capabilities Capabilities
 }
 
 func New() *Flags {