+++ pkg/graph/resolver/resolver.go
@@ -21,6 +21,7 @@ type ResolverConfig struct {
 	InstallationK8sApiUrl  string
 	InstallationK8sAuthUrl string
 	InstallationK8sCaCert  string
+	AvailabilityZones      []string
 }
 
 type Resolver struct {
@@ -31,6 +32,7 @@ type Resolver struct {
 	installationK8sApiUrl  string
 	installationK8sAuthUrl string
 	installationK8sCaCert  string
+	availabilityZones      []string
 }
 
 func NewResolver(config ResolverConfig) (*Resolver, error) {
@@ -45,6 +47,7 @@ func NewResolver(config ResolverConfig) (*Resolver, error) {
 		installationK8sApiUrl:  formatUrl(config.InstallationK8sApiUrl),
 		installationK8sAuthUrl: formatUrl(config.InstallationK8sAuthUrl),
 		installationK8sCaCert:  certificate.Parse(config.InstallationK8sCaCert),
+		availabilityZones:      config.AvailabilityZones,
 	}
 
 	return r, nil